third_party/blink/web_tests/fast/dom/shadow/host-context-pseudo-class-css-text.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<style id="style1">
.foo:host-context(body.mytheme) > div { display: block; }
:host-context(*) { display: block; }
</style>
<style id="style-invalid">
:host-context(.a, .b) { display: block }
:host-context(div div) { display: block }
:host-context(div > div) { display: block }
:host-context(div + div) { display: block }
:host-context(div ~ div) { display: block }
:host-context) { display: block }
:host-context() { display: block; }
</style>
</head>
<script>
description("Test for cssText of ':host-context()' rule.");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(0).cssText", ".foo:host-context(body.mytheme) > div { display: block; }");
shouldBeEqualToString("document.getElementById('style1').sheet.cssRules.item(1).cssText", ":host-context(*) { display: block; }");
shouldBe("document.getElementById('style-invalid').sheet.cssRules.length", "0");
</script>
</html>
